Profiling (#94)
* initial test suite

* small compiler changes

* fixing found executor bugs

* fixed most tests

* more executor bug fixes

* satisified with findings

+++ src/Buckle/Compiler/CodeAnalysis/Binding/ConstantFolding.cs
@@ -89,7 +89,7 @@ internal static class ConstantFolding {
                 if (specialType == SpecialType.Int)
                     return new ConstantValue((long)Math.Pow((long)leftValue, (long)rightValue), specialType);
                 else
-                    return new ConstantValue((double)Math.Pow((double)leftValue, (double)rightValue), specialType);
+                    return new ConstantValue(Math.Pow((double)leftValue, (double)rightValue), specialType);
             case BinaryOperatorKind.ConditionalAnd:
                 return new ConstantValue((bool)leftValue && (bool)rightValue, specialType);
             case BinaryOperatorKind.ConditionalOr:
